glucoguide-ai-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-multipart
pydantic
firebase-admin
langchain
langchain-core
langchain-groq
pypdf
gunicorn
psycopg2-binary
